Cardano.YTxP.SDK.ControlParameters

Description

Deprecated: Use Ply instead.

This module is consider deprecated. Use Ply instead.

Types

data YieldingScripts

Scripts that yield to transaction families identified by reference scripts which carry a state thread token

Constructors

YieldingScripts

Fields

yieldingMintingPolicies :: [HexStringScript "YieldingMP"]

Since: 0.1.0

yieldingValidator :: HexStringScript "YieldingValidator"

Since: 0.1.0

yieldingStakingValidators :: [HexStringScript "YieldingSV"]

Since: 0.1.0

data ControlParameters

Contains the compiled scripts along with the parameters they were compiled against. This is useful for _library consumers_ and should contain all of the information needed to work with the library.

Constructors

ControlParameters

Fields

yieldingScripts :: YieldingScripts

Since: 0.1.0

sdkParameters :: SdkParameters

Since: 0.1.0

newtype HexStringScript (scriptLabel :: Symbol)

A helper newtype to ensure that any Scripts we use are serialized (and deserialized), consistently. We also have a 'tag' for a more specific type name or label when using a via derivation with this, or as part of a larger type.

See JSON.md for an explanation of our policy. TODO: update json policy

Constructors

HexStringScript ShortByteString

sbsToHexText :: ShortByteString -> Text

Converts a ShortByteString into a textual representation in hex.

hexTextToSbs :: MonadFail m => Text -> m ShortByteString

Attempts to parse the given Text into the ShortByteString that would have produced it via sbsToHexText, indicating parse failures with fail.