Cardano.YTxP.SDK.Redeemers

Documentation

newtype AuthorisedScriptIndex

Represents an index into a authorised reference script in a TxInReferenceInput list

Constructors

AuthorisedScriptIndex Integer

data AuthorisedScriptPurpose

Constructors

Minting

Spending

Rewarding

newtype AuthorisedScriptProofIndex

Constructors

AuthorisedScriptProofIndex (AuthorisedScriptPurpose, Integer)

data YieldingRedeemer

The redeemer passed to the yielding minting policy, validator, and staking validators

Constructors

YieldingRedeemer

Fields

authorisedScriptIndex :: AuthorisedScriptIndex

Integer The index of the TxInReferenceInput that contains the authorised reference script.

authorisedScriptProofIndex :: AuthorisedScriptProofIndex

A tuple containing yielded to script type and the index at which to find proof: this allows us to avoid having to loop through inputsmintswithdrawls to find the script we want to ensure is run.