Cardano.YTxP.SDK.SdkParameters

Description

Documentation

data SdkParameters

Parameters available during compilation (therefore not containing any script hashes).

Constructors

SdkParameters

Fields

stakingValidatorsNonceList :: [Natural]

A list of nonces for the yielding staking validators. One staking validator is compiled for each nonce. @since 0.1.0

mintingPoliciesNonceList :: [Natural]

A list of nonces for the yielding minting policies. One minting policy is compiled for each nonce. @since 0.1.0

certifyingValidatorsNonceList :: [Natural]

A list of nonces for the certifying staking validators. One certifying validator is compiled for each nonce. @since 0.2.0

votingValidatorsNonceList :: [Natural]

A list of nonces for the voting staking validators. One voting validator is compiled for each nonce. @since 0.2.0

proposingValidatorsNonceList :: [Natural]

A list of nonces for the proposing staking validators. One proposing validator is compiled for each nonce. @since 0.2.0

authorisedScriptsSTCS :: AuthorisedScriptsSTCS

The Currency symbol of the token that identifies authorised reference scripts . @since 0.1.0

newtype AuthorisedScriptsSTCS

Semantic newtype for the YieldList state thread currency symbol

Constructors

AuthorisedScriptsSTCS CurrencySymbol